+++ powershell-module/Az.Tools.Migration/Tests/Functions/Public/Find-AzUpgradeCommandReference.tests.ps1
@@ -6,6 +6,8 @@ InModuleScope -ModuleName Az.Tools.Migration -ScriptBlock {
             # arrange
             Mock -CommandName Test-Path -MockWith { Write-Output -InputObject $true }
 
+            Mock -CommandName Resolve-Path -MockWith { @{Path = $Path} }
+
             # ensure we don't send telemetry during tests.
             Mock -CommandName Send-MetricsIfDataCollectionEnabled -ModuleName Az.Tools.Migration -MockWith { }
 
@@ -71,7 +73,7 @@ InModuleScope -ModuleName Az.Tools.Migration -ScriptBlock {
             }
 
             # act
-            $moduleSpec = Get-AzUpgradeCmdletSpec -ModuleName "AzureRM" -ModuleVersion "6.13.1"
+            $moduleSpec = Get-AzUpgradeCmdletSpec -AzureRM
             $results = Find-AzUpgradeCommandReference -FilePath "testfile.ps1" -AzureRmModuleSpec $moduleSpec
 
             # assert
@@ -132,7 +134,7 @@ InModuleScope -ModuleName Az.Tools.Migration -ScriptBlock {
             }
 
             # act
-            $moduleSpec = Get-AzUpgradeCmdletSpec -ModuleName "AzureRM" -ModuleVersion "6.13.1"
+            $moduleSpec = Get-AzUpgradeCmdletSpec -AzureRM
             $results = Find-AzUpgradeCommandReference -DirectoryPath "C:\test" -AzureRmModuleSpec $moduleSpec
 
             # assert