+++ powershell-module/Az.Tools.Migration/Tests/Functions/Public/Get-AzUpgradeAliasSpec.tests.ps1
@@ -2,10 +2,10 @@ Import-Module Az.Tools.Migration -Force
 
 InModuleScope -ModuleName Az.Tools.Migration -ScriptBlock {
     Describe 'Get-AzUpgradeAliasSpec tests' {
-        It 'Should be able to import the Az 9.3.0 alias mapping spec' {
+        It 'Should be able to import the Az alias mapping spec' {
             # arrange/act
-            $expectedAliasCount = 2623
-            $spec = Get-AzUpgradeAliasSpec -ModuleVersion "9.3.0"
+            $expectedAliasCount = 2626
+            $spec = Get-AzUpgradeAliasSpec -ModuleVersion latest
 
             # assert
             $spec | Should Not Be $null