+++ vscode-extension/src/powershell.ts
@@ -24,19 +24,24 @@ export class PowershellProcess {
         });
     }
 
-    //exec the migration command and get the result
-    public async getUpgradePlan(filePath: string, azureRmVersion: string, azVersion: string): Promise<string> {
-        //const command = `New-AzUpgradeModulePlan -FilePath "${filePath}" -FromAzureRmVersion "${azureRmVersion}" -ToAzVersion "${azVersion}" | ConvertTo-Json -depth 10`;
+    //exec the migration command and get the result to latest Az version
+    public async getUpgradePlanToLatest(filePath: string, azureRmVersion: string): Promise<string> {
         if (this.powershell.invocationStateInfo == "Running") {
             //the latter cancels the former powershell process
             await this.restart();
         }
 
-        const command = `New-AzUpgradeModulePlan -FilePath "${filePath}" -FromAzureRmVersion "${azureRmVersion}" -ToAzVersion "${azVersion}" | ConvertTo-Json`;
+        const command = `New-AzUpgradeModulePlan -FilePath "${filePath}" -FromAzureRmVersion "${azureRmVersion}" -ToAzVersion "latest" | ConvertTo-Json`;
         this.powershell.addCommand(command);
         const planResult = await this.powershell.invoke();
 
         return planResult;
+    }
+
+    //exec the migration command and get the result to LTS Az version
+    public async getUpgradePlanToLTS(filePath: string, azureRmVersion: string): Promise<string> {
+        //TODO: Implement this in the future when LTS ready
+        return;
     }
 
     //check whether the module exists
@@ -47,6 +52,37 @@ export class PowershellProcess {
             moduleFolder => fs.existsSync(path.resolve(moduleFolder, moduleName))
         );
 
+    }
+
+    //check the version of the installed module
+    public checkModuleVersion(moduleName: string): string {
+        const systemModulePath = this.getSystemModulePath();
+        let moduleFolder = systemModulePath.find(
+            moduleFolder => fs.existsSync(path.resolve(moduleFolder, moduleName))
+        );
+        moduleFolder = path.join(moduleFolder, moduleName);
+        const versions = fs.readdirSync(moduleFolder);
+        let latest = versions[0];
+        for (let i = 1; i < versions.length; i++) {
+            if (this.versionCompare(versions[i], latest) > 0) {
+                latest = versions[i];
+            }
+        }
+        return latest;
+    }
+
+    //TODO: handle more edge cases, e.g. -preview
+    public versionCompare(v1, v2): number {
+        v1 = v1.split('.');
+        v2 = v2.split('.');
+        const k = Math.min(v1.length, v2.length);
+        for (let i = 0; i < k; i++) {
+            v1[i] = parseInt(v1[i], 10);
+            v2[i] = parseInt(v2[i], 10);
+            if (v1[i] > v2[i]) return 1;
+            if (v1[i] < v2[i]) return -1;        
+        }
+        return v1.length == v2.length ? 0: (v1.length < v2.length ? -1 : 1);
     }
 
     //install the module automatically